coverity: fix defects
- 360250 Macro compares unsigned to 0
- 360278 Macro compares unsigned to 0
- 360280 Macro compares unsigned to 0
- 360284 Macro compares unsigned to 0
- 360321 Macro compares unsigned to 0
- 488556 Dereference null return value

+++ copy.c
@@ -74,7 +74,7 @@ ARRAY_HEAD(HeaderArray, char *);
  * If a header already exists in that position, the new text will be
  * concatenated on the old.
  */
-static void add_one_header(struct HeaderArray *headers, size_t pos, char *value)
+static void add_one_header(struct HeaderArray *headers, int pos, char *value)
 {
   char **old = ARRAY_GET(headers, pos);
   if (old && *old)
